Alg Name, matchup selection indications
The algorithm tables have the corresponding name over them.
The type matchup list now indicates which matchup is active. The title
name above the list is removed as it was not used. The matchup list
indicates the current by showing it as green and disabling the button.
Visual fix of the algorithms checkboxes having the same width. Algorithm
tables are seperated with black line.