Changed threads to multiprocessing module.
The use of threads was changed to use the multiprocessing module as it
supports killing threads easier.
Threads now are killed if they take to long to complete.
The mutable pass_info dictionary was changed to a multiprocessing Queue
as multiprocessing does not support sharing data with mutable objects,
but does support using a Queue.