Support Resizing Chart and Results Views.
The views of the type chart and the algorithm results support being
resized automatically based on the root window size. The scrollbar is
used to control the view. A canvas holds the overall frame to enable
scrolling. This was achieved with row/column configure and setting the
weight to a positive number for the cell of the canvas. The canvas is
then added NSEW sticky to the parent container frame.
The minimum size of the window is set using .after so that the size is
properly accounting for width of each widget. The minimum size is
configured using the selections of matchups and algorithms, and
assuming a minimum size of 400 by 300 for the chart and algorithm
results.